chore: add separate `git_chdir` option to RubyGit::CommandLine.run
Using the `git_chdir: <path>` option will pass the `-C <path>` option to the
git command.  This will cause git to run as if it was started in <path>
instead of the current working directory.

This is necesary to use this instead of the `chdir: <path>` option which get
passed to `Process::Spawn` because JRuby does not support the `chdir` option.