refactor: revert changes to make path normalization optional
Revert the changes introduced in #64 that allowed the user of this lib to make
path normalization optional. The purpose of #64 was to make tests easier to
write. The downside was that the library code itself became more complicated.

Instead this change removes that and adds mocking for path normalization
when it is appropriate to shut it off for tests.